Stop basic agent crashing in certain conditions
BasicAgent would crash when the following conditions were met
* A BasicAgent decides to give a random hint
* It is given to an agent in the final round, and that agent does not have a full hand due to no remaining deck

The crash is a NullPointerException (or whatever java calls it) when it attempts to get the value or colour of the missing card in the hand being hinted at in order to create the array of booleans for `Action`.
It looks like this was fixed in `hint` but not `hintRandom`.

I will also note that giving this hint is probably pointless, as the only time an agent is missing a card is after their final turn and as such can't use the hint.